requirements.txt
pytube
SpeechRecognition
pydub
youtube_dl
python-ffmpeg
python-dotenv
assemblyai>=0.15
langchain==0.2.14
langchain-community==0.2.12
langchain-groq===0.1.9
google-generativeai
sentence-transformers
moviepy
requests
ffmpeg-python
deepgram-sdk==2.12.0
git+https://github.com/m-bain/whisperx.git
pyannote.audio
streamlit
fastapi
python-multipart
uvicorn
boto3
